art-aggregator/requirements.txt
spotipy
python-dotenv
pymongo
requests
flask
flask-restful
jq
ecs-logging
pyyaml
flask-cors
pytest
coverage
pytest-cov
